Derive.Derive

Description

Samodejno računanje vrednosti funkcije in njenih odvodov v dani točki.

Documentation

data D a

Vrednost funkcije in njene odvode v neki točki bomo predstavili s tipom D.

Constructors

D a (Maybe (D a))

constD :: Num a => a -> D a

constD je konstanta.

idD :: Num a => a -> D a

idD je identiteta.

sqr :: Num a => a -> a

first_n :: Floating a => (D a -> D a) -> a -> Int -> [a]

Funkcija first_n nam izpiše vrednost funkcije f ter prvih n odvodov v neki točki x.

nth :: Floating a => (D a -> D a) -> a -> Int -> a

Funkcija nth nam izpiše n-ti odvod funkcije f v točki x.